updated technology page design
Dean suggested to separate the "What is a skill sprint" section by a fullwidth block of color, but because we have everything inside a container to make sure the content is aligned on different screens this idea is not possible to implement. Plus, I didn't like that section in a color bloc. This page needs to be looked at more